Moves annual goal to its own file
The user model file is just,, very long
