+++ bookwyrm/models/__init__.py
@@ -17,7 +17,8 @@ from .attachment import Image
 from .favorite import Favorite
 from .readthrough import ReadThrough, ProgressUpdate, ProgressMode
 
-from .user import User, KeyPair, AnnualGoal
+from .user import User, KeyPair
+from .annual_goal import AnnualGoal
 from .relationship import UserFollows, UserFollowRequest, UserBlocks
 from .report import Report, ReportComment
 from .federated_server import FederatedServer
