+++ bookwyrm/models/user.py
@@ -5,7 +5,6 @@ from urllib.parse import urlparse
 from django.apps import apps
 from django.contrib.auth.models import AbstractUser, Group
 from django.contrib.postgres.fields import ArrayField, CICharField
-from django.core.validators import MinValueValidator
 from django.dispatch import receiver
 from django.db import models, transaction
 from django.utils import timezone
@@ -16,7 +15,7 @@ import pytz
 from bookwyrm import activitypub
 from bookwyrm.connectors import get_data, ConnectorException
 from bookwyrm.models.shelf import Shelf
-from bookwyrm.models.status import Status, Review
+from bookwyrm.models.status import Status
 from bookwyrm.preview_images import generate_user_preview_image_task
 from bookwyrm.settings import DOMAIN, ENABLE_PREVIEW_IMAGES, USE_HTTPS, LANGUAGES
 from bookwyrm.signatures import create_key_pair
@@ -25,7 +24,7 @@ from bookwyrm.utils import regex
 from .activitypub_mixin import OrderedCollectionPageMixin, ActivitypubMixin
 from .base_model import BookWyrmModel, DeactivationReason, new_access_code
 from .federated_server import FederatedServer
-from . import fields, Review
+from . import fields
 
 
 FeedFilterChoices = [
@@ -412,65 +411,6 @@ class KeyPair(ActivitypubMixin, BookWyrmModel):
         if not self.public_key:
             self.private_key, self.public_key = create_key_pair()
         return super().save(*args, **kwargs)
-
-
-def get_current_year():
-    """sets default year for annual goal to this year"""
-    return timezone.now().year
-
-
-class AnnualGoal(BookWyrmModel):
-    """set a goal for how many books you read in a year"""
-
-    user = models.ForeignKey("User", on_delete=models.PROTECT)
-    goal = models.IntegerField(validators=[MinValueValidator(1)])
-    year = models.IntegerField(default=get_current_year)
-    privacy = models.CharField(
-        max_length=255, default="public", choices=fields.PrivacyLevels
-    )
-
-    class Meta:
-        """unqiueness constraint"""
-
-        unique_together = ("user", "year")
-
-    def get_remote_id(self):
-        """put the year in the path"""
-        return f"{self.user.remote_id}/goal/{self.year}"
-
-    @property
-    def books(self):
-        """the books you've read this year"""
-        return (
-            self.user.readthrough_set.filter(
-                finish_date__year__gte=self.year,
-                finish_date__year__lt=self.year + 1,
-            )
-            .order_by("-finish_date")
-            .all()
-        )
-
-    @property
-    def ratings(self):
-        """ratings for books read this year"""
-        book_ids = [r.book.id for r in self.books]
-        reviews = Review.objects.filter(
-            user=self.user,
-            book__in=book_ids,
-        )
-        return {r.book.id: r.rating for r in reviews}
-
-    @property
-    def progress(self):
-        """how many books you've read this year"""
-        count = self.user.readthrough_set.filter(
-            finish_date__year__gte=self.year,
-            finish_date__year__lt=self.year + 1,
-        ).count()
-        return {
-            "count": count,
-            "percent": int(float(count / self.goal) * 100),
-        }
 
 
 @app.task(queue="low_priority")
